FIX: Ordering of fields (WIP)
The GraphQL spec says the order of the fields in the output should match
the order specified in the query.

Maps in Elixir do not provide a way to traverse the keys in source
order and maps are currently used to hold fields in query definitions
and during execution.

This means order of fields in the output is non-deterministic right now.

The fix for this is to use keyword lists instead of maps in query
definitions and during execution.

A further complication is that Poison (used in the plug) cannot encode
keyword lists, so we'd need to have some alternative solution for
transforming the literal query results into JSON.