Implement the AST Reducer
The AST Reducer walks the AST of a GraphQL query and invokes a Visitor
as nodes of the AST are traversed. The Visitor is responsible for
maintaining an accumulator.

At the end of the traversal, the final values in the accumulator can be
obtained. This lays the groundwork for implementing query validators
(the majority, if not all, will be Visitor implementations).