Provide means to combine multiple Visitors into one
This will allow us to write a type information visitor that 'wraps' a
validation visitor so that when *enter* and *leave* are called on the
validation visitor, type information for the current node will be
available on the accumulator.

It also makes it easy to insert visitors in a chain and debug what's
going on when writing tests.