Implement the TypeInfoVisitor
This injects type information into the accumulator for subsequent
visitors (that have been combined using the CompositeVisitor) have
access to the type information of the current AST node.

NOTE: this is complex and there are no tests!

The implementation is translated from the JS version (which also has no
direct tests) but it tested indirectly via the validation tests. I plan
on taking the same approach.